Sometimes slack returns 200 but with an inner error, we need to handle this
